+++ client/version/main.go
@@ -1,3 +1,3 @@
 // This file is auto-generated
 package version
-var Version string = "0.8.0-rc.1"
+var Version string = "0.8.0"