+++ config/services/watcher.mjs
@@ -9,7 +9,7 @@ export const pullConfig = {
   // Image to run
   image: 'docker.elastic.co/beats/heartbeat',
   // Image tag (version) to run
-  tag: '8.15.3',
+  tag: '8.16.5',
 }
 
 /*