+++ scripts/pull-oci-images.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
-docker pull docker.redpanda.com/redpandadata/redpanda:v24.2.7
-docker pull smallstep/step-ca:0.27.5
-docker pull timberio/vector:0.44.0-debian
-docker pull docker.redpanda.com/redpandadata/console:v2.7.2
-docker pull rqlite/rqlite:8.32.3
-docker pull traefik:v3.1.6
-docker pull docker.elastic.co/beats/heartbeat:8.15.3
+docker pull docker.redpanda.com/redpandadata/redpanda:v24.2.20
+docker pull smallstep/step-ca:0.28.2
+docker pull timberio/vector:0.45.0-debian
+docker pull docker.redpanda.com/redpandadata/console:v2.8.3
+docker pull rqlite/rqlite:8.36.12
+docker pull traefik:v3.3.4
+docker pull docker.elastic.co/beats/heartbeat:8.16.5