+++ ui/pages/actions/clients/enroll/index.mjs
@@ -47,8 +47,8 @@ const EnrollClientsPage = (props) => {
         <Popout note>
           <h5>This cluster requires client invites</h5>
           <p>
-            The <code>REQUIRE_CLIENT_INVITES</code> feature flag is enabled on this cluster,
-            meaning that clients cannot join without providing an invite code.
+            The <code>REQUIRE_CLIENT_INVITES</code> feature flag is enabled on this cluster, meaning
+            that clients cannot join without providing an invite code.
           </p>
           <p>
             If you have an invite code, you can enter it below to auto-update the command examples.
@@ -115,7 +115,7 @@ const EnrollClientsPage = (props) => {
                 To install the Morio client for Linux, run the following command from an account
                 with <code>sudo</code> rights:
               </p>
-              <Highlight language="shell">{`curl -fsSL https://install.morio.it/client/ | bash`}</Highlight>
+              <Highlight language="shell">{`curl -fsSL https://install.morio.it/client | bash`}</Highlight>
             </Tab>
             <Tab name="macOS" key="mac">
               <Popout fixme>Provide instructions for macOS</Popout>